requirements.txt
asgiref==3.8.1
boto3==1.35.21
botocore==1.35.21
dj-database-url==0.5.0
Django==3.2.25
django-allauth==0.55.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
s3transfer==0.10.2
sqlparse==0.5.1
stripe==10.10.0
